pyproject.toml
[project]
name = "blazeplot"
version = "0.1.0"
description = "BlazePlot listens to a local Fire Dispatch frequency and plots all of the updates come from it on a map to make it more visual."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "keyboard>=0.13.5",
    "numpy>=2.4.3",
    "openai-whisper>=20250625",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
]

[build-system]
requires = ["uv_build>=0.9.17,<0.10.9"]
build-backend = "uv_build"

[tool.setuptools]
package-dir = { "" = "src" }
